# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JCEmpleos API",
        "description": "Documentación oficial del API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/perfil/idioma": {
            "get": {
                "tags": [
                    "Perfil - Idiomas"
                ],
                "summary": "Listar idiomas del usuario autenticado",
                "operationId": "eb4d7b772ca140fde4d247fd34b8bec3",
                "responses": {
                    "200": {
                        "description": "Idiomas obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perfil - Idiomas"
                ],
                "summary": "Agregar idioma al perfil",
                "operationId": "1b908eeff8b7fb21c330348506bf0e38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idioma_id",
                                    "nivel_id"
                                ],
                                "properties": {
                                    "idioma_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nivel_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Idioma agregado correctamente"
                    },
                    "422": {
                        "description": "Error al agregar idioma"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perfil/idioma/{id}": {
            "put": {
                "tags": [
                    "Perfil - Idiomas"
                ],
                "summary": "Actualizar nivel de idioma",
                "operationId": "c5ab8507e4fd61f17537d2c5b1541386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nivel_id"
                                ],
                                "properties": {
                                    "nivel_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Idioma actualizado correctamente"
                    },
                    "422": {
                        "description": "Error al actualizar idioma"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perfil - Idiomas"
                ],
                "summary": "Eliminar idioma del perfil",
                "operationId": "2117440bb241983fb93c733e711fcc4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Idioma eliminado correctamente"
                    },
                    "422": {
                        "description": "Error al eliminar idioma"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuarios/{id}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "operationId": "00be41590d60944dab4a27f179022793",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "email",
                                    "estado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@email.com"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente"
                    },
                    "422": {
                        "description": "Error al actualizar usuario"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuarios/{id}/desactivar": {
            "patch": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Desactivar usuario (soft delete)",
                "operationId": "4c93eba72e388265ea49948913a2f82c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a desactivar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario desactivado correctamente"
                    },
                    "422": {
                        "description": "Error al desactivar usuario"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}